third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-empty-srcset-in-main-frame-blocked-expected.txt

CONSOLE WARNING: line 2: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-empty-srcset.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/security/resources/compass.jpg': this content should also be loaded over HTTPS.

This test opens a window that loads an insecure image with a srcset attribute.
